+++ examples/motoko_examples/minimal-counter-dapp/canisters/azle/index.ts
