+++ examples/motoko_examples/minimal-counter-dapp/canisters/declarations/azle/azle.did
@@ -1,5 +1,5 @@
 service: {
+    "getCount": () -> (nat) query;
     "count": () -> (nat);
-    "getCount": () -> (nat);
     "reset": () -> (nat);
 }
