+++ examples/motoko_examples/minimal-counter-dapp/canisters/declarations/azle/azle.did.d.ts
@@ -1,0 +1,8 @@
+import type { Principal } from '@dfinity/principal';
+import type { ActorMethod } from '@dfinity/agent';
+
+export interface _SERVICE {
+    count: ActorMethod<[], bigint>;
+    getCount: ActorMethod<[], bigint>;
+    reset: ActorMethod<[], bigint>;
+}
