+++ examples/motoko_examples/minimal-counter-dapp/canisters/declarations/azle/azle.did.js
@@ -1,7 +1,7 @@
 export const idlFactory = ({ IDL }) => {
     return IDL.Service({
         count: IDL.Func([], [IDL.Nat], []),
-        getCount: IDL.Func([], [IDL.Nat], []),
+        getCount: IDL.Func([], [IDL.Nat], ['query']),
         reset: IDL.Func([], [IDL.Nat], [])
     });
 };
