+++ examples/motoko_examples/minimal-counter-dapp/canisters/declarations/minimal_dapp_assets/index.js
