+++ examples/motoko_examples/minimal-counter-dapp/canisters/declarations/minimal_dapp_assets/minimal_dapp_assets.did.js
