+++ examples/motoko_examples/minimal-counter-dapp/canisters/minimal_dapp_assets/assets/sample-asset.txt
