+++ examples/motoko_examples/minimal-counter-dapp/canisters/minimal_dapp_assets/src/index.html
