+++ examples/motoko_examples/minimal-counter-dapp/canisters/minimal_dapp_assets/src/index.js
@@ -1,4 +1,4 @@
-import { minimal_dapp } from '../../declarations/minimal_dapp';
+import { azle as minimal_dapp } from '../../declarations/azle';
 
 document.addEventListener('DOMContentLoaded', async function () {
     const counter = await minimal_dapp.getCount();
