+++ examples/motoko_examples/minimal-counter-dapp/dfx_generated/azle/azle.did
