+++ examples/motoko_examples/minimal-counter-dapp/package.json
@@ -14,7 +14,7 @@
         "prebuild": "npm run copy:types",
         "start": "webpack serve --mode development --env development",
         "prestart": "npm run copy:types",
-        "copy:types": "rsync -avr .dfx/$(echo ${DFX_NETWORK:-'**'})/canisters/** --exclude='assets/' --exclude='idl/' --exclude='*.wasm' --delete src/declarations"
+        "copy:types": "rsync -avr .dfx/$(echo ${DFX_NETWORK:-'**'})/canisters/** --exclude='assets/' --exclude='idl/' --exclude='*.wasm' --delete canisters/declarations"
     },
     "dependencies": {
         "azle": "0.5.0"
@@ -25,7 +25,6 @@
         "@dfinity/principal": "0.10.3",
         "assert": "2.0.0",
         "buffer": "6.0.3",
-        "copy-webpack-plugin": "^9.0.1",
         "events": "3.3.0",
         "html-webpack-plugin": "5.5.0",
         "process": "0.11.10",
