--- examples/motoko_examples/minimal-counter-dapp/test/dfx_generated/minimal_dapp/minimal_dapp.did.d.ts
@@ -1,6 +1,0 @@
-import type { Principal } from '@dfinity/principal';
-export interface _SERVICE {
-    count: () => Promise<bigint>;
-    getCount: () => Promise<bigint>;
-    reset: () => Promise<bigint>;
-}
